requirements.txt
fireworks-ai
python-dotenv==1.0.1
PyYAML==6.0.2

